refactor(test): Move PipelineBuilder to orca-api-test
